+++ ouvICEx/init_flask.py
@@ -1,10 +1,10 @@
 from flask import Flask, render_template
 
-from database.posts import db
-from admin.posts import admin_view
+from scripts.database import db
+from scripts.admin_login import admin_login
 
 app = Flask(__name__)
-app.register_blueprint(admin_view, url_prefix="/admin")
+app.register_blueprint(admin_login, url_prefix="/admin")
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///posts.sqlite3'
 app.config['SQLALCHEMY_TRACK_NOTIFICATIONS'] = False
