+++ ouvICEx/scripts/templates/admin_login.html
@@ -1,0 +1,6 @@
+{% extends 'base.html' %}
+{% block title %} Login {% endblock %}
+
+{% block content %}
+    <h1> Login </h1>
+{% endblock %}
